+++ Goose/Commands/HairdyeCommand.cs
@@ -80,7 +80,7 @@ namespace Goose.Commands
                             "" + "," + // Guild name
                             prevx + "," +
                             prevy + "," +
-                            ctx.Player.Facing + "," +
+                            (int)ctx.Player.Facing + "," +
                             100 + "," + // HP %
                             ctx.Player.BodyID + "," +
                             ctx.Player.BodyR + "," + // Body Color R
@@ -111,7 +111,7 @@ namespace Goose.Commands
                             "" + "," + // Guild name
                             prevx + "," +
                             prevy + "," +
-                            ctx.Player.Facing + "," +
+                            (int)ctx.Player.Facing + "," +
                             100 + "," + // HP %
                             ctx.Player.BodyID + "," +
                             ctx.Player.BodyR + "," + // Body Color R
